- updated BookRental to use collectionView
